feat: update amount field to accept float values and adjust related components for decimal handling

+++ Antonieta.Backend/dtos/create_distribution_dto.py
@@ -7,5 +7,5 @@ class create_distribution_dto(BaseModel):
     """
     ration_id: int
     beneficiary_id: int
-    amount: int
+    amount: float
     date: datetime
